[#113][product] refactor: dto 관련 보완
